+++ src/odemis/util/dc_fetch.py
@@ -22,11 +22,10 @@ Odemis. If not, see http://www.gnu.org/licenses/.
 Retrieval helpers for downloading DataCollector ZIP samples from S3.
 """
 
-import argparse
 import logging
 from datetime import datetime, timezone
 from pathlib import Path
-from typing import Any, Dict, Iterator, List, Optional
+from typing import Any, Dict, Iterator, List, Optional, Tuple
 
 from odemis.util.datacollector import DataCollectorConfig, S3UploadBackend
 
@@ -84,66 +83,6 @@ def parse_key_event_name(key: str) -> Optional[str]:
     if len(parts) != 3:
         return None
     return parts[0] or None
-
-def build_argument_parser() -> argparse.ArgumentParser:
-    """
-    Build CLI argument parser for odemis-dc-fetch.
-    :return: Configured ArgumentParser instance.
-    """
-    examples = (
-        "Examples:\n"
-        "  odemis-dc-fetch\n"
-        "  odemis-dc-fetch --output ./downloads\n"
-        "  odemis-dc-fetch --event z_stack_acquired\n"
-        "  odemis-dc-fetch --since 2026-03-01\n"
-        "  odemis-dc-fetch --host meteor-5099\n"
-        "  odemis-dc-fetch --host meteor-5099,atlas-001,secom-22\n"
-        "  odemis-dc-fetch --bucket delmic-odemis-collect-test --region eu-west-1\n"
-        "  odemis-dc-fetch --since 2026-03-01T12:30:00 --event z_stack_acquired --output ./dc_samples"
-    )
-    parser = argparse.ArgumentParser(
-        description="Fetch data-collection ZIP samples from S3.",
-        epilog=examples,
-        formatter_class=argparse.RawDescriptionHelpFormatter,
-    )
-    parser.add_argument(
-        "--event",
-        dest="event",
-        help="Only fetch samples matching event name.",
-    )
-    parser.add_argument(
-        "--since",
-        dest="since",
-        help="Only fetch samples since UTC date/datetime (e.g. 2026-03-01 or 2026-03-01T12:30:00).",
-    )
-    parser.add_argument(
-        "--output",
-        dest="output",
-        default="./dc_samples",
-        help="Output directory for downloaded ZIPs (default: ./dc_samples).",
-    )
-    parser.add_argument(
-        "--host",
-        dest="host",
-        help="Optional host/system-id filter; use comma-separated IDs for multiple hosts. "
-             "By default, fetch across all hosts.",
-    )
-    parser.add_argument(
-        "--bucket",
-        dest="bucket",
-        help="Optional S3 bucket override (default from datacollector backend config).",
-    )
-    parser.add_argument(
-        "--endpoint-url",
-        dest="endpoint_url",
-        help="Optional S3 endpoint URL override (default from datacollector backend config).",
-    )
-    parser.add_argument(
-        "--region",
-        dest="region",
-        help="Optional AWS region name for S3 client creation.",
-    )
-    return parser
 
 def iter_s3_objects(s3_client: Any, bucket: str, prefix: str) -> Iterator[Dict[str, Any]]:
     """
@@ -206,7 +145,7 @@ def build_s3_client_from_config(
     bucket_override: Optional[str] = None,
     endpoint_override: Optional[str] = None,
     region_override: Optional[str] = None,
-) -> tuple[Any, str]:
+) -> Tuple[Any, str]:
     """
     Build an S3 client using datacollector credentials with optional endpoint/bucket overrides.
     :param config: DataCollectorConfig instance.
@@ -304,35 +243,3 @@ def fetch_samples(
         "skipped_existing": skipped_existing,
         "failed": failed,
     }
-
-def main(argv: Optional[List[str]] = None) -> int:
-    """CLI entrypoint for retrieval flow."""
-    parser = build_argument_parser()
-    args = parser.parse_args(argv)
-
-    try:
-        since_utc = parse_since_utc(args.since) if args.since else None
-    except ValueError:
-        logging.error("Invalid --since value: %s", args.since)
-        return 2
-
-    output_dir = Path(args.output)
-    try:
-        result = fetch_samples(
-            event_filter=args.event,
-            since_utc=since_utc,
-            output_dir=output_dir,
-            host_filter=args.host,
-            bucket_override=args.bucket,
-            endpoint_override=args.endpoint_url,
-            region_override=args.region,
-        )
-    except Exception:
-        logging.exception("Failed to fetch samples from S3")
-        return 1
-
-    print(
-        "listed={listed} matched={matched} downloaded={downloaded} "
-        "skipped_existing={skipped_existing} failed={failed}".format(**result)
-    )
-    return 0 if result["failed"] == 0 else 1
